replace downloader and add parameter
