+++ cotk/_utils/file_utils.py
@@ -12,6 +12,7 @@ import hashlib
 import logging
 import sys
 from pathlib import Path
+from urllib.parse import urlparse
 import requests
 
 from tqdm import tqdm
@@ -267,3 +268,48 @@ def import_local_resources(file_id, local_path, cache_dir=CACHE_DIR, \
 		return cache_path
 	else:
 		raise ValueError("bad hashtag of {}".format(res_name))
+
+def load_file_from_url(url, force=False, cache_dir=CACHE_DIR):
+	'''Download a file from the given ``url``. If the file has been downloaded, it will be cached in ``cache_dir``.
+	However, the function can't check whether the file from ``url`` is changed online. If you have the wrong cache,
+	you may manually delete the file cached located by the return value.
+
+	Arguments:
+		url(str): A url indicating the file online.
+		force(bool): Force to download and ignore the existing file.
+		cache_dir(str, option): A path indicating where the cache place. If ``None``, a default cache path is
+			used.
+
+	Returns:
+		(str) The local path of downloaded model.
+
+	Example:
+		>>> load_model_from_url('https://s3.amazonaws.com/pytorch/models/resnet18-5c106cde.pth')
+		>>> {CACHE_DIR}/resnet18-5c106cde.pth
+	'''
+
+	parts = urlparse(url)
+	filename = os.path.basename(parts.path)
+
+	cache_dir = os.path.join(cache_dir, 'models')
+
+	if os.path.exists(cache_dir) and force:
+		shutil.rmtree(cache_dir)
+		# raise ValueError("model existed. If you want to delete the existing model. \
+		# 	Use `rm %s`." % cache_path)
+
+	os.makedirs(cache_dir, exist_ok=True)
+	cache_path = os.path.join(cache_dir, filename)
+	if os.path.exists(cache_path):
+		return cache_path
+
+	with tempfile.NamedTemporaryFile() as temp_file:
+		_http_get(url, temp_file)
+		temp_file.flush() # flush to avoid truncation
+		temp_file.seek(0) # shutil.copyfileobj() starts at the current position
+
+		with open(cache_path, 'wb') as cache_file:
+			shutil.copyfileobj(temp_file, cache_file)
+
+	LOGGER.info('model cached at %s', cache_path)
+	return cache_path
