+++ tests/_utils/test_file_loadmodel.py
@@ -5,7 +5,7 @@ import shutil
 import os
 
 from cotk._utils.file_utils import _get_file_sha256
-from cotk._utils.downloader import load_model_from_url
+from cotk.downloader import load_file_from_url
 
 @pytest.fixture
 def r_mock():
@@ -16,13 +16,13 @@ class TestFileUtils():
 	def test_MSCOCO_resource(self):
 		cache_dir = './tests/_utils/dataset_cache'
 		data_dir = './tests/_utils/data'
-		res_path = load_model_from_url('https://cotk-data.s3-ap-northeast-1.amazonaws.com/test.zip', cache_dir=cache_dir)
+		res_path = load_file_from_url('https://cotk-data.s3-ap-northeast-1.amazonaws.com/test.zip', cache_dir=cache_dir)
 
 		with pytest.raises(ValueError) as excinfo:
-			load_model_from_url('https://cotk-data.s3-ap-northeast-1.amazonaws.com/test.zip', cache_dir=cache_dir)
-		assert("model existed. If you want to delete the existing model." in str(excinfo.value))
+			load_file_from_url('https://cotk-data.s3-ap-northeast-1.amazonaws.com/test.zip', cache_dir=cache_dir)
+		assert "model existed. If you want to delete the existing model." in str(excinfo.value)
 
-		assert(res_path == os.path.join(cache_dir, 'models', 'test.zip'))
-		assert(_get_file_sha256(res_path) == _get_file_sha256(os.path.join(data_dir, 'test.zip')))
+		assert res_path == os.path.join(cache_dir, 'models', 'test.zip')
+		assert _get_file_sha256(res_path) == _get_file_sha256(os.path.join(data_dir, 'test.zip'))
 
 		shutil.rmtree(cache_dir)
